2018-5-18
修订了管理员资料更改，
主页菜单数据展示，
员工产品按管理员查看

+++ src/com/huan/business/action/UserAction.java
@@ -38,7 +38,7 @@ public class UserAction extends ActionSupport implements ModelDriven<TsUser> {
 	 */
 	public String getupUser() {
 		rolesName = this.roleService.getRoleNameById(user.getManageId());
-		TsUser luser = this.userService.getUserById(user.getUserId().intValue());
+		TsUser luser = this.userService.getUserById(user.getUserId());
 		powerList =this.powerService.getPowerList();
 		if (luser!= null) {
 			user = luser;
@@ -86,7 +86,7 @@ public class UserAction extends ActionSupport implements ModelDriven<TsUser> {
 		}
 	}
 	public String getInfoUser() {
-		user = userService.getUserById(user.getUserId().intValue());
+		user = userService.getUserById(user.getUserId());
 		user.setManageName(userService.getManageById(user.getManageId().intValue()).getManageName());
 		return "userInfo";
 	}
