+++ src/com/huan/business/service/IUserService.java
@@ -1,5 +1,7 @@
 package com.huan.business.service;
 
+import java.math.BigDecimal;
+
 import com.huan.business.po.TsManage;
 import com.huan.business.po.TsUser;
 import com.huan.tool.PageBean;
@@ -10,7 +12,7 @@ public interface IUserService {
 	public Boolean addUser(TsUser user);
 	public Boolean updateUser(TsUser user);
 	public Boolean delUser(TsUser user);
-	public TsUser getUserById(int userId);
+	public TsUser getUserById(BigDecimal userId);
 	public Boolean delUsers(String str);
 	public TsUser getUserByName(String userName);
 	public String getUserNameById(int userid);
